CustomGPT/requirements.txt
# Core AI/ML Libraries
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-anthropic>=0.1.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Document Processing
pypdf2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0

# Google Drive Integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Web Scraping & APIs
requests>=2.31.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.2

# Database & Storage
sqlalchemy>=2.0.0

# UI & Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.0.0

# Email & Communication
email-validator>=2.0.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
